# docs/openapi/repository-api-v1.json
{
  "components": {
    "schemas": {
      "ChecksumResponse": {
        "description": "Checksum metadata attached to stored content or versions.",
        "properties": {
          "algorithm": {
            "description": "Checksum algorithm used by the service.",
            "title": "Algorithm",
            "type": "string"
          },
          "digest": {
            "description": "Hex digest returned by the service.",
            "title": "Digest",
            "type": "string"
          }
        },
        "required": [
          "algorithm",
          "digest"
        ],
        "title": "ChecksumResponse",
        "type": "object"
      },
...
      "ErrorBody": {
        "description": "Error detail object for the public API error envelope.",
        "properties": {
          "code": {
            "description": "Stable machine-readable error code.",
            "title": "Code",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional structured metadata to help clients debug the error.",
            "title": "Details"
          },
          "message": {
            "description": "Human-readable summary of the failure.",
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorEnvelope": {
        "description": "Standardized error envelope shared across HTTP endpoints.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody",
            "description": "Normalized error payload."
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
      "ProvenanceResponse": {
        "description": "Minimal provenance returned by immutable version reads.",
        "properties": {
          "commit_sha": {
            "title": "Commit Sha",
            "type": "string"
          },
          "repo_url": {
            "title": "Repo Url",
            "type": "string"
          },
          "tree_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tree Path"
          }
        },
        "required": [
          "repo_url",
          "commit_sha"
        ],
        "title": "ProvenanceResponse",
        "type": "object"
      },
...
      "SkillContentSummaryResponse": {
        "description": "Compact content metadata returned without the full markdown body.",
        "properties": {
          "checksum": {
            "$ref": "#/components/schemas/ChecksumResponse"
          },
          "rendered_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional pre-rendered short summary stored with the markdown.",
            "title": "Rendered Summary"
          },
          "size_bytes": {
            "description": "UTF-8 byte length of the stored markdown.",
            "title": "Size Bytes",
            "type": "integer"
          }
        },
        "required": [
          "checksum",
          "size_bytes",
          "rendered_summary"
        ],
        "title": "SkillContentSummaryResponse",
        "type": "object"
      },
...
      "SkillGovernanceRequest": {
        "additionalProperties": false,
        "description": "Governance metadata supplied at publish time.",
        "properties": {
          "provenance": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProvenanceRequest"
              },
              {
                "type": "null"
              }
            ]
          },
          "trust_tier": {
            "default": "untrusted",
            "enum": [
              "untrusted",
              "internal",
              "verified"
            ],
            "title": "Trust Tier",
            "type": "string"
          }
        },
        "title": "SkillGovernanceRequest",
        "type": "object"
      },
      "SkillMetadataResponse": {
        "description": "Full normalized metadata block returned by immutable metadata reads.",
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "headers": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Headers"
          },
          "inputs_schema": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Inputs Schema"
          },
          "maturity_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Maturity Score"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "outputs_schema": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Outputs Schema"
          },
          "security_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Security Score"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          },
          "token_estimate": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Token Estimate"
          }
        },
        "required": [
          "name",
          "description",
          "tags"
        ],
        "title": "SkillMetadataResponse",
        "type": "object"
      },
...
      "SkillVersionContentRequest": {
        "additionalProperties": false,
        "description": "Markdown body provided at publish time.",
        "properties": {
          "raw_markdown": {
            "description": "Canonical markdown body for this immutable version.",
            "title": "Raw Markdown",
            "type": "string"
          },
          "rendered_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional pre-rendered short summary stored alongside the markdown.",
            "title": "Rendered Summary"
          }
        },
        "required": [
          "raw_markdown"
        ],
        "title": "SkillVersionContentRequest",
        "type": "object"
      },
...
      "SkillVersionCreateRequest": {
        "additionalProperties": false,
        "description": "Normalized JSON publish contract.",
        "properties": {
          "content": {
            "$ref": "#/components/schemas/SkillVersionContentRequest"
          },
          "governance": {
            "$ref": "#/components/schemas/SkillGovernanceRequest"
          },
          "metadata": {
            "$ref": "#/components/schemas/SkillVersionMetadataRequest"
          },
          "relationships": {
            "$ref": "#/components/schemas/SkillVersionRelationshipsRequest"
          },
          "slug": {
            "description": "Stable public slug for the skill identity.",
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9](?:[A-Za-z0-9._-]{0,127})$",
            "title": "Slug",
            "type": "string"
          },
          "version": {
            "description": "Immutable semantic version being published.",
            "pattern": "^(0|[1-9]\\d*)\\.(0|[1-9]\\d*)\\.(0|[1-9]\\d*)(?:-([0-9A-Za-z-]+(?:\\.[0-9A-Za-z-]+)*))?(?:\\+([0-9A-Za-z-]+(?:\\.[0-9A-Za-z-]+)*))?$",
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "slug",
          "version",
          "content",
          "metadata"
        ],
        "title": "SkillVersionCreateRequest",
        "type": "object"
      },
...
      "SkillVersionMetadataRequest": {
        "additionalProperties": false,
        "description": "Structured query metadata provided at publish time.",
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional human-readable summary of the skill.",
            "title": "Description"
          },
          "headers": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "Flexible header-like attributes stored as JSON.",
            "title": "Headers"
          },
          "inputs_schema": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "Structured input contract stored as JSON.",
            "title": "Inputs Schema"
          },
          "maturity_score": {
            "anyOf": [
              {
                "maximum": 1.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Normalized maturity score in the range [0, 1].",
            "title": "Maturity Score"
          },
          "name": {
            "description": "Human-readable skill name.",
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "outputs_schema": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "Structured output contract stored as JSON.",
            "title": "Outputs Schema"
          },
          "security_score": {
            "anyOf": [
              {
                "maximum": 1.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Normalized security score in the range [0, 1].",
            "title": "Security Score"
          },
          "tags": {
            "description": "Free-form tags used for categorization and discovery.",
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          },
          "token_estimate": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Approximate token footprint for ranking/filtering.",
            "title": "Token Estimate"
          }
        },
        "required": [
          "name"
        ],
        "title": "SkillVersionMetadataRequest",
        "type": "object"
      },
      "SkillVersionMetadataResponse": {
        "description": "Immutable metadata envelope returned by publish and metadata batch fetch.",
        "properties": {
          "content": {
            "$ref": "#/components/schemas/SkillContentSummaryResponse"
          },
          "lifecycle_status": {
            "enum": [
              "published",
              "deprecated",
              "archived"
            ],
            "title": "Lifecycle Status",
            "type": "string"
          },
          "metadata": {
            "$ref": "#/components/schemas/SkillMetadataResponse"
          },
          "provenance": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProvenanceResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "published_at": {
            "format": "date-time",
            "title": "Published At",
            "type": "string"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "trust_tier": {
            "enum": [
              "untrusted",
              "internal",
              "verified"
            ],
            "title": "Trust Tier",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "string"
          },
          "version_checksum": {
            "$ref": "#/components/schemas/ChecksumResponse"
          }
        },
        "required": [
          "slug",
          "version",
          "version_checksum",
          "content",
          "metadata",
          "lifecycle_status",
          "trust_tier",
          "published_at"
        ],
        "title": "SkillVersionMetadataResponse",
        "type": "object"
      },
      "SkillVersionRelationshipsRequest": {
        "additionalProperties": false,
        "description": "Grouped authored relationships provided at publish time.",
        "properties": {
          "conflicts_with": {
            "items": {
              "$ref": "#/components/schemas/ExactRelationshipSelectorRequest"
            },
            "title": "Conflicts With",
            "type": "array"
          },
          "depends_on": {
            "items": {
              "$ref": "#/components/schemas/DependencySelectorRequest"
            },
            "title": "Depends On",
            "type": "array"
          },
          "extends": {
            "items": {
              "$ref": "#/components/schemas/ExactRelationshipSelectorRequest"
            },
            "title": "Extends",
            "type": "array"
          },
          "overlaps_with": {
            "items": {
              "$ref": "#/components/schemas/ExactRelationshipSelectorRequest"
            },
            "title": "Overlaps With",
            "type": "array"
          }
        },
        "title": "SkillVersionRelationshipsRequest",
        "type": "object"
      },
      "SkillVersionStatusResponse": {
        "description": "Lifecycle status update response.",
        "properties": {
          "is_current_default": {
            "title": "Is Current Default",
            "type": "boolean"
          },
          "lifecycle_changed_at": {
            "format": "date-time",
            "title": "Lifecycle Changed At",
            "type": "string"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "status": {
            "enum": [
              "published",
              "deprecated",
              "archived"
            ],
            "title": "Status",
            "type": "string"
          },
          "trust_tier": {
            "enum": [
              "untrusted",
              "internal",
              "verified"
            ],
            "title": "Trust Tier",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "slug",
          "version",
          "status",
          "trust_tier",
          "lifecycle_changed_at",
          "is_current_default"
        ],
        "title": "SkillVersionStatusResponse",
        "type": "object"
      },
      "SkillVersionStatusUpdateRequest": {
        "additionalProperties": false,
        "description": "Lifecycle transition request for one immutable version.",
        "properties": {
          "note": {
            "anyOf": [
              {
                "maxLength": 1000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "status": {
            "enum": [
              "published",
              "deprecated",
              "archived"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "SkillVersionStatusUpdateRequest",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Registry-first API for immutable skill publication, candidate discovery,\nexact dependency reads, immutable metadata batch fetch, immutable content batch\nfetch, and governed lifecycle updates.\n\nThe server owns data-local registry operations only. Prompt interpretation,\nreranking, dependency solving, final selection, lock generation, and execution\nplanning remain client-owned behavior outside this API boundary.",
    "title": "aptitude-server Service",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/skill-versions": {
      "post": {
        "description": "Create a new immutable skill version from a normalized JSON payload containing markdown content, structured metadata, and authored relationships.",
        "operationId": "createSkillVersion",
        "requestBody": {
          "content": {
            "application/json": {
              "example": {
                "content": {
                  "raw_markdown": "# Python Lint\n\nLint Python files consistently.\n",
                  "rendered_summary": "Lint Python files consistently."
                },
                "governance": {
                  "provenance": {
                    "commit_sha": "aabbccddeeff00112233445566778899aabbccdd",
                    "repo_url": "https://github.com/example/skills",
                    "tree_path": "skills/python.lint"
                  },
                  "trust_tier": "internal"
                },
                "metadata": {
                  "description": "Linting skill",
                  "headers": {
                    "runtime": "python"
                  },
                  "inputs_schema": {
                    "type": "object"
                  },
                  "maturity_score": 0.9,
                  "name": "Python Lint",
                  "outputs_schema": {
                    "type": "object"
                  },
                  "security_score": 0.95,
                  "tags": [
                    "python",
                    "lint"
                  ],
                  "token_estimate": 128
                },
                "relationships": {
                  "conflicts_with": [],
                  "depends_on": [
                    {
                      "markers": [
                        "linux",
                        "gpu"
                      ],
                      "optional": true,
                      "slug": "python.base",
                      "version_constraint": ">=1.0.0,<2.0.0"
                    }
                  ],
                  "extends": [
                    {
                      "slug": "python.base",
                      "version": "1.0.0"
                    }
                  ],
                  "overlaps_with": [
                    {
                      "slug": "python.format",
                      "version": "1.0.0"
                    }
                  ]
                },
                "slug": "python.lint",
                "version": "1.2.3"
              },
              "schema": {
                "$ref": "#/components/schemas/SkillVersionCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "example": {
                  "content": {
                    "checksum": {
                      "algorithm": "sha256",
                      "digest": "c3ab8ff13720e8ad9047dd39466b3c8974e592c2fa383d4a3960714caef0c4f2"
                    },
                    "rendered_summary": "Lint Python files consistently.",
                    "size_bytes": 123
                  },
                  "lifecycle_status": "published",
                  "metadata": {
                    "description": "Linting skill",
                    "headers": {
                      "runtime": "python"
                    },
                    "inputs_schema": {
                      "type": "object"
                    },
                    "maturity_score": 0.9,
                    "name": "Python Lint",
                    "outputs_schema": {
                      "type": "object"
                    },
                    "security_score": 0.95,
                    "tags": [
                      "python",
                      "lint"
                    ],
                    "token_estimate": 128
                  },
                  "provenance": {
                    "commit_sha": "aabbccddeeff00112233445566778899aabbccdd",
                    "repo_url": "https://github.com/example/skills",
                    "tree_path": "skills/python.lint"
                  },
                  "published_at": "2026-03-10T08:30:00Z",
                  "slug": "python.lint",
                  "trust_tier": "internal",
                  "version": "1.2.3",
                  "version_checksum": {
                    "algorithm": "sha256",
                    "digest": "c3ab8ff13720e8ad9047dd39466b3c8974e592c2fa383d4a3960714caef0c4f2"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/SkillVersionMetadataResponse"
                }
              }
            },
            "description": "Immutable skill version published successfully."
          },
          "409": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "DUPLICATE_SKILL_VERSION",
                    "details": {
                      "slug": "python.lint",
                      "version": "1.2.3"
                    },
                    "message": "Skill version already exists: python.lint@1.2.3"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The requested immutable `slug@version` already exists."
          },
          "422": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "INVALID_REQUEST",
                    "details": {
                      "errors": [
                        {
                          "input": "latest",
                          "loc": [
                            "path",
                            "version"
                          ],
                          "msg": "String should match pattern",
                          "type": "string_pattern_mismatch"
                        }
                      ]
                    },
                    "message": "Request validation failed."
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The request body, path parameters, or query parameters are invalid."
          },
          "500": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "CONTENT_STORAGE_FAILURE",
                    "message": "Failed to persist immutable skill version."
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Normalized content or metadata persistence failed."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Publish an immutable skill version",
        "tags": [
          "skills"
        ]
      }
    },
    "/skills/{slug}/versions/{version}/status": {
      "patch": {
        "description": "Transition the lifecycle state for one immutable skill version.",
        "operationId": "updateSkillVersionStatus",
        "parameters": [
          {
            "description": "Stable public slug of the skill.",
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "description": "Stable public slug of the skill.",
              "pattern": "^[A-Za-z0-9](?:[A-Za-z0-9._-]{0,127})$",
              "title": "Slug",
              "type": "string"
            }
          },
          {
            "description": "Exact immutable semantic version to update.",
            "in": "path",
            "name": "version",
            "required": true,
            "schema": {
              "description": "Exact immutable semantic version to update.",
              "pattern": "^(0|[1-9]\\d*)\\.(0|[1-9]\\d*)\\.(0|[1-9]\\d*)(?:-([0-9A-Za-z-]+(?:\\.[0-9A-Za-z-]+)*))?(?:\\+([0-9A-Za-z-]+(?:\\.[0-9A-Za-z-]+)*))?$",
              "title": "Version",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SkillVersionStatusUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "is_current_default": true,
                  "lifecycle_changed_at": "2026-03-11T09:15:00Z",
                  "slug": "python.lint",
                  "status": "deprecated",
                  "trust_tier": "internal",
                  "version": "1.2.3"
                },
                "schema": {
                  "$ref": "#/components/schemas/SkillVersionStatusResponse"
                }
              }
            },
            "description": "Lifecycle status updated successfully."
          },
          "404": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "SKILL_VERSION_NOT_FOUND",
                    "details": {
                      "slug": "python.lint",
                      "version": "9.9.9"
                    },
                    "message": "Skill version not found: python.lint@9.9.9"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The requested immutable `slug@version` does not exist."
          },
          "422": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "INVALID_REQUEST",
                    "details": {
                      "errors": [
                        {
                          "input": "latest",
                          "loc": [
                            "path",
                            "version"
                          ],
                          "msg": "String should match pattern",
                          "type": "string_pattern_mismatch"
                        }
                      ]
                    },
                    "message": "Request validation failed."
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The request body, path parameters, or query parameters are invalid."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update immutable version lifecycle status",
        "tags": [
          "skills"
        ]
      }
    }
  }
}